# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskFlow API",
        "description": "API REST de gestão de tarefas para times com suporte a webhooks.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://taskflow-api.test",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/workspaces/{workspace}/members": {
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Convidar membro",
                "operationId": "9d713e2b353006c982cd6e4fcdd8450b",
                "parameters": [
                    {
                        "name": "workspace",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "joao@test.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Membro adicionado"
                    },
                    "422": {
                        "description": "Já é membro"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/workspaces/{workspace}/members/{user}": {
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Remover membro",
                "operationId": "a50b7799250d9a6e4b8e695eb329b385",
                "parameters": [
                    {
                        "name": "workspace",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}